requirements.txt
flask==2.3.3
httpx
streamlit
openai
python-dotenv
agents-sdk